refactor: 식별자 생성 클래스 패키지 위치 변경 (#45)
