+++ src/main/kotlin/kr/wooco/woocobe/auth/infrastructure/storage/AuthUserEntity.kt
@@ -9,7 +9,7 @@ import jakarta.persistence.Table
 import kr.wooco.woocobe.auth.domain.model.AuthUser
 import kr.wooco.woocobe.auth.domain.model.SocialAuthInfo
 import kr.wooco.woocobe.auth.domain.model.SocialAuthType
-import kr.wooco.woocobe.common.domain.IdGenerator
+import kr.wooco.woocobe.common.infrastructure.storage.IdGenerator
 import kr.wooco.woocobe.common.storage.BaseTimeEntity
 
 @Entity
