+++ src/main/kotlin/kr/wooco/woocobe/common/infrastructure/storage/IdGenerator.kt
@@ -1,10 +1,10 @@
-package kr.wooco.woocobe.common.domain
+package kr.wooco.woocobe.common.infrastructure.storage
 
 import com.github.f4b6a3.tsid.TsidFactory
-import kr.wooco.woocobe.common.config.IdGeneratorConfig
 
 object IdGenerator {
-    private val tsidFactory: TsidFactory = TsidFactory.builder()
+    private val tsidFactory: TsidFactory = TsidFactory
+        .builder()
         .withNode(IdGeneratorConfig.node)
         .build()
