+++ src/main/kotlin/kr/wooco/woocobe/user/infrastructure/storage/UserEntity.kt
@@ -4,7 +4,7 @@ import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.domain.IdGenerator
+import kr.wooco.woocobe.common.infrastructure.storage.IdGenerator
 import kr.wooco.woocobe.common.storage.BaseTimeEntity
 import kr.wooco.woocobe.user.domain.model.User
